NTLN-227: fixed possible bug which nstimeinterval is treated as a float value
git-svn-id: http://www.physalis.net/repos/natsulion/trunk@405 5e439dd1-9381-4ddc-8459-9b31d6a87cac

+++ src/NTLNConfiguration.m
@@ -148,7 +148,7 @@ static id<NTLNTimelineSortOrderChangeObserver> _timelineSortOrderChangeObserver;
 }
 
 - (void) setLatestTimestampOfMessage:(NSTimeInterval)interval {
-    [[NSUserDefaults standardUserDefaults] setFloat:interval forKey:@"latestTimestampOfMessage"];
+    [[NSUserDefaults standardUserDefaults] setDouble:interval forKey:@"latestTimestampOfMessage"];
 }
 
 - (NSTimeInterval) latestTimestampOfDirectMessage {
@@ -156,7 +156,7 @@ static id<NTLNTimelineSortOrderChangeObserver> _timelineSortOrderChangeObserver;
 }
 
 - (void) setLatestTimestampOfDirectMessage:(NSTimeInterval)interval {
-    [[NSUserDefaults standardUserDefaults] setFloat:interval forKey:@"latestTimestampOfDirectMessage"];
+    [[NSUserDefaults standardUserDefaults] setDouble:interval forKey:@"latestTimestampOfDirectMessage"];
 }
 
 - (BOOL) useGrowlAndSummarizeGrowl {
